src/services/yolo/requirements.txt
Flask==2.3.3
torch==2.0.1
torchvision==0.15.2
ultralytics==8.0.120
opencv-python-headless==4.8.0.74
pillow==10.0.0
numpy==1.24.3
